+++ src/lib/analytics.js
@@ -1,17 +1,20 @@
 const FuzzySet = require('fuzzyset');
 
+const { getS3Object } = require('../lib/s3');
+
 const DEBUG = process.env.DEBUG || false;
 const DURATION_VARIANCE_PERCENTAGE = Number.parseFloat(process.env.DURATION_VARIANCE_PERCENTAGE);
-const FUZZY_SET_INVALID_MATCHES = JSON.parse(process.env.FUZZY_SET_INVALID_MATCHES);
+const FORCED_MATCHES = process.env.FORCED_MATCHES;
+const FUZZY_SET_INVALID_MATCHES = process.env.FUZZY_SET_INVALID_MATCHES;
 const FUZZY_SET_MINIMUM_SCORE_MATCH = Number.parseFloat(process.env.FUZZY_SET_MINIMUM_SCORE_MATCH);
-const LECTURES_SKIP_LIST = JSON.parse(process.env.LECTURES_SKIP_LIST);
+const SKIPPED_MATCHES = process.env.SKIPPED_MATCHES;
 
-const analyzeCourseSharedContent = (course, otherCourses) => {
+const analyzeCourseSharedContent = (course, otherCourses, configuration) => {
   for (const otherCourse of otherCourses) {
     const fuzzyset = FuzzySet(Object.keys(otherCourse.lectures), false);
 
     for (const lectureTitle of Object.keys(course.lectures)) {
-      if (LECTURES_SKIP_LIST.includes(course.lectures[lectureTitle].titleWithDuration)) {
+      if (configuration.skippedMatches.includes(course.lectures[lectureTitle].titleWithDuration)) {
         continue;
       }
 
@@ -41,16 +44,12 @@ const analyzeCourseSharedContent = (course, otherCourses) => {
               lecture = course.lectures[lectureTitle];
               otherLecture = otherCourse.lectures[result[1]];
 
-              validFuzzyMatchFound = isValidMatch(lecture, otherLecture);
+              validFuzzyMatchFound = isValidMatch(lecture, otherLecture, configuration);
 
               if (validFuzzyMatchFound) {
                 break;
               }
             }
-          }
-
-          if (!validFuzzyMatchFound) {
-            continue;
           }
         }
 
@@ -90,17 +89,74 @@ const analyzeCourseSharedContent = (course, otherCourses) => {
             otherLecture.sharedWith[course.code] = lecture.titleWithDuration;
           }
         }
+
+        if (!(otherCourse.code in course.lectures[lectureTitle].sharedWith)) {
+          if (course.lectures[lectureTitle].titleWithDuration.toLowerCase() in configuration.forcedMatches) {
+            for (const otherLectureTitle of Object.keys(otherCourse.lectures)) {
+              if (
+                configuration.forcedMatches[course.lectures[lectureTitle].titleWithDuration.toLowerCase()].includes(
+                  otherCourse.lectures[otherLectureTitle].titleWithDuration.toLowerCase()
+                )
+              ) {
+                lecture = course.lectures[lectureTitle];
+                otherLecture = otherCourse.lectures[otherLectureTitle];
+
+                lecture.sharedWith[otherCourse.code] = otherLecture.titleWithDuration;
+                otherLecture.sharedWith[course.code] = lecture.titleWithDuration;
+
+                console.log(
+                  `Forced Match\n${course.code} => ${lecture.titleWithDuration}\n${otherCourse.code} => ${
+                    otherLecture.titleWithDuration
+                  }\n${'='.repeat(80)}`
+                );
+
+                break;
+              }
+            }
+          }
+        }
       }
     }
   }
 };
 
-const isValidMatch = (lecture, otherLecture) => {
+const initializeConfiguration = async () => {
+  const forcedMatches = JSON.parse(await getS3Object(FORCED_MATCHES));
+
+  for (const titleWithDuration of Object.keys(forcedMatches)) {
+    forcedMatches[titleWithDuration.toLowerCase()] = forcedMatches[titleWithDuration].map(
+      (forcedMatchingTitleWithDuration) => {
+        return forcedMatchingTitleWithDuration.toLowerCase();
+      }
+    );
+    delete forcedMatches[titleWithDuration];
+  }
+
+  for (const [titleWithDuration, forcedMatchingTitlesWtihDuration] of Object.entries(forcedMatches)) {
+    for (const forcedMatchingTitleWtihDuration of forcedMatchingTitlesWtihDuration) {
+      const lowerCaseForcedMatchingTitleWtihDuration = forcedMatchingTitleWtihDuration.toLowerCase();
+
+      if (lowerCaseForcedMatchingTitleWtihDuration in forcedMatches) {
+        forcedMatches[lowerCaseForcedMatchingTitleWtihDuration].push(titleWithDuration);
+      } else {
+        forcedMatches[lowerCaseForcedMatchingTitleWtihDuration] = [titleWithDuration];
+      }
+    }
+  }
+
+  return {
+    forcedMatches: forcedMatches,
+    fuzzySetInvalidMatches: JSON.parse(await getS3Object(FUZZY_SET_INVALID_MATCHES)),
+    skippedMatches: JSON.parse(await getS3Object(SKIPPED_MATCHES))
+  };
+};
+
+const isValidMatch = (lecture, otherLecture, configuration) => {
   if (
-    (lecture.titleWithDuration in FUZZY_SET_INVALID_MATCHES &&
-      FUZZY_SET_INVALID_MATCHES[lecture.titleWithDuration].includes(otherLecture.titleWithDuration)) ||
-    (otherLecture.titleWithDuration in FUZZY_SET_INVALID_MATCHES &&
-      FUZZY_SET_INVALID_MATCHES[otherLecture.titleWithDuration].includes(lecture.titleWithDuration))
+    (lecture.titleWithDuration in configuration.fuzzySetInvalidMatches &&
+      configuration.fuzzySetInvalidMatches[lecture.titleWithDuration].includes(otherLecture.titleWithDuration)) ||
+    (otherLecture.titleWithDuration in configuration.fuzzySetInvalidMatches &&
+      configuration.fuzzySetInvalidMatches[otherLecture.titleWithDuration].includes(lecture.titleWithDuration))
   ) {
     return false;
   }
@@ -108,10 +164,12 @@ const isValidMatch = (lecture, otherLecture) => {
   return true;
 };
 
-exports.analyzeCourses = (courses) => {
+exports.analyzeCourses = async (courses) => {
+  const configuration = await initializeConfiguration();
+
   const flattenedCourses = Object.values(courses).flat();
 
   for (let i = 0; i < flattenedCourses.length; i++) {
-    analyzeCourseSharedContent(flattenedCourses[i], flattenedCourses.slice(i + 1));
+    analyzeCourseSharedContent(flattenedCourses[i], flattenedCourses.slice(i + 1), configuration);
   }
 };